requirements.txt
Django==1.5.1
MySQL-python==1.2.4
argparse==1.2.1
beautifulsoup4==4.2.1
distribute==0.6.34
django-registration==1.0
psycopg2==2.5.1
wsgiref==0.1.2
